+++ WildAnalyticsSDK/WildAnalyticsSDKTests/Mocks/EventsProcessorMock.swift
@@ -23,6 +23,7 @@ final class EventsProcessorMock: EventsProcessor {
     private(set) var setupWasCalled: Int = 0
     private(set) var setupReceivedApiKey: String?
     private(set) var setupReceivedIsFirstLaunch: Bool?
+    private(set) var setupReceivedEnableAutomaticEvents: Bool?
     private(set) var setupReceivedDropCache: Bool?
     private(set) var setupReceivedQueue: DispatchQueue?
     private(set) var setupReceivedBatchConfig: BatchConfig?
@@ -37,6 +38,7 @@ final class EventsProcessorMock: EventsProcessor {
     func setup(
         apiKey: String,
         isFirstLaunch: Bool,
+        enableAutomaticEvents: Bool,
         dropCache: Bool,
         queue: DispatchQueue?,
         batchConfig: BatchConfig,
@@ -49,6 +51,7 @@ final class EventsProcessorMock: EventsProcessor {
         setupWasCalled += 1
         setupReceivedApiKey = apiKey
         setupReceivedIsFirstLaunch = isFirstLaunch
+        setupReceivedEnableAutomaticEvents = enableAutomaticEvents
         setupReceivedDropCache = dropCache
         setupReceivedQueue = queue
         setupReceivedBatchConfig = batchConfig